// projects/choice-gas/src/app/data/enum/settings.enum.ts
export enum Setting {
	Charges = 'charges',
	EnabledHeatmaps = 'enabledHeatmaps',
	EnabledOfferTypes = 'enabledOfferTypes',
	EnableHighDensityTable = 'highDensity',
	EnableStripedTable = 'stripedTables',
	EnrollmentFields = 'enrollmentFields',
	OfferRateOverrides = 'rateOverrides',
	Scheme = 'scheme',
	SeriesOverrides = 'seriesOverrides',
	ShowColumns = 'showColumns',
	ShowSeries = 'showSeries',
}